+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/WalletService.kt
@@ -5,10 +5,7 @@ import at.asitplus.catching
 import at.asitplus.openid.*
 import at.asitplus.openid.OpenIdConstants.Errors
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
-import at.asitplus.signum.indispensable.josef.JsonWebKeySet
-import at.asitplus.signum.indispensable.josef.JsonWebToken
-import at.asitplus.signum.indispensable.josef.JwsHeader
-import at.asitplus.signum.indispensable.josef.toJwsAlgorithm
+import at.asitplus.signum.indispensable.josef.*
 import at.asitplus.wallet.lib.agent.CryptoService
 import at.asitplus.wallet.lib.agent.DefaultCryptoService
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
@@ -34,6 +31,8 @@ import io.ktor.util.*
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.datetime.Clock
 import kotlin.random.Random
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.minutes
 
 /**
  * Client service to retrieve credentials using OID4VCI
@@ -361,3 +360,84 @@ suspend fun JwsService.buildDPoPHeader(
     addJsonWebKey = true,
     addX5c = false,
 ).getOrThrow().serialize()
+
+/**
+ * Client attestation JWT, issued by the backend service to a client, which can be sent to an OAuth2 Authorization
+ * Server if needed, e.g. as HTTP header `OAuth-Client-Attestation`, see
+ * [OAuth 2.0 Attestation-Based Client Authentication](https://www.ietf.org/archive/id/draft-ietf-oauth-attestation-based-client-auth-04.html)
+ *
+ * @param clientId OAuth 2.0 client ID of the wallet
+ * @param issuer a unique identifier for the entity that issued the JWT
+ * @param clientKey key to be attested, i.e. included in a [ConfirmationClaim]
+ * @param keyType optional key type acc. to OID4VC HAIP with SD-JWT VC to include in the [ConfirmationClaim]
+ * @param userAuthentication optional user authentication acc. to OID4VC HAIP with SD-JWT VC to include in the [ConfirmationClaim]
+ * @param lifetime validity period of the assertion (minus the [clockSkew])
+ * @param clockSkew duration to subtract from [Clock.System.now] when setting the creation timestamp
+ */
+suspend fun JwsService.buildClientAttestationJwt(
+    clientId: String,
+    issuer: String,
+    clientKey: JsonWebKey,
+    keyType: WalletAttestationKeyType? = null,
+    userAuthentication: WalletAttestationUserAuthentication? = null,
+    authenticationLevel: String? = null,
+    lifetime: Duration = 60.minutes,
+    clockSkew: Duration = 5.minutes,
+) = createSignedJwsAddingParams(
+    header = JwsHeader(
+        algorithm = algorithm,
+        type = JwsContentTypeConstants.CLIENT_ATTESTATION_JWT
+    ),
+    payload = JsonWebToken(
+        issuer = issuer,
+        subject = clientId,
+        issuedAt = Clock.System.now() - clockSkew,
+        expiration = Clock.System.now() - clockSkew + lifetime,
+        authenticationLevel = authenticationLevel,
+        confirmationClaim = ConfirmationClaim(
+            jsonWebKey = clientKey,
+            keyType = keyType,
+            userAuthentication = userAuthentication,
+        )
+    ),
+    serializer = JsonWebToken.serializer(),
+    addKeyId = false,
+    addJsonWebKey = false,
+    addX5c = false,
+).getOrThrow()
+
+/**
+ * Client attestation PoP JWT, issued by the client, which can be sent to an OAuth2 Authorization Server if needed,
+ * e.g. as HTTP header `OAuth-Client-Attestation-PoP`, see
+ * [OAuth 2.0 Attestation-Based Client Authentication](https://www.ietf.org/archive/id/draft-ietf-oauth-attestation-based-client-auth-04.html)
+ *
+ * @param clientId OAuth 2.0 client ID of the wallet
+ * @param audience The RFC8414 issuer identifier URL of the authorization server MUST be used
+ * @param nonce optionally provided from the authorization server
+ * @param lifetime validity period of the assertion (minus the [clockSkew])
+ * @param clockSkew duration to subtract from [Clock.System.now] when setting the creation timestamp
+ */
+suspend fun JwsService.buildClientAttestationPoPJwt(
+    clientId: String,
+    audience: String,
+    nonce: String? = null,
+    lifetime: Duration = 10.minutes,
+    clockSkew: Duration = 5.minutes
+) = createSignedJwsAddingParams(
+    header = JwsHeader(
+        algorithm = algorithm,
+        type = JwsContentTypeConstants.CLIENT_ATTESTATION_POP_JWT
+    ),
+    payload = JsonWebToken(
+        issuer = clientId,
+        audience = audience,
+        jwtId = Random.nextBytes(12).encodeToString(Base64UrlStrict),
+        nonce = nonce,
+        issuedAt = Clock.System.now() - clockSkew,
+        expiration = Clock.System.now() - clockSkew + lifetime,
+    ),
+    serializer = JsonWebToken.serializer(),
+    addKeyId = false,
+    addJsonWebKey = false,
+    addX5c = false,
+).getOrThrow()